# docs/07_interfaces/openapi_v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevPilot Local API Contract v1",
    "version": "1.0.0-implemented-initial",
    "description": "Static OpenAPI contract synchronized with the implemented-initial local FastAPI MVP. FUNC-SPRINT-67 implements the server adapter on localhost; Sprint 68 will add token/CORS hardening."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Future local-only server. Not active in Sprint 66."
    }
  ],
...
  "paths": {
...
    "/api/v1/validation/frontmatter": {
      "post": {
        "operationId": "validation_frontmatter",
        "summary": "API-VALIDATION-FRONTMATTER: validation.frontmatter",
        "description": "Contract-only endpoint mapped to `ValidationApplicationService.validate_frontmatter`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "validation"
        ],
        "x-devpilot-api-id": "API-VALIDATION-FRONTMATTER",
        "x-devpilot-operation": "validation.frontmatter",
        "x-devpilot-domain-service": "ValidationApplicationService.validate_frontmatter",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "token-planned-sprint-68",
        "x-devpilot-policy": "PathGuard on adapter",
        "x-devpilot-command-equivalent": "python -m devpilot_core validate-frontmatter <path> --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplicationRequest"
                  }
                ],
                "properties": {
                  "operation": {
                    "const": "validation.frontmatter"
                  }
                }
              },
              "example": {
                "operation": "validation.frontmatter",
                "payload": {},
                "client": "api-local",
                "dry_run": true
              }
            }
          }
        }
      }
    },
    "/api/v1/validation/artifact": {
      "post": {
        "operationId": "validation_artifact",
        "summary": "API-VALIDATION-ARTIFACT: validation.artifact",
        "description": "Contract-only endpoint mapped to `ValidationApplicationService.validate_artifact`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "validation"
        ],
        "x-devpilot-api-id": "API-VALIDATION-ARTIFACT",
        "x-devpilot-operation": "validation.artifact",
        "x-devpilot-domain-service": "ValidationApplicationService.validate_artifact",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "none",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "token-planned-sprint-68",
        "x-devpilot-policy": "PathGuard on adapter",
        "x-devpilot-command-equivalent": "python -m devpilot_core validate-artifact <path> --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplicationRequest"
                  }
                ],
                "properties": {
                  "operation": {
                    "const": "validation.artifact"
                  }
                }
              },
              "example": {
                "operation": "validation.artifact",
                "payload": {},
                "client": "api-local",
                "dry_run": true
              }
            }
          }
        }
      }
    },
    "/api/v1/validation/readiness": {
      "post": {
        "operationId": "validation_readiness",
        "summary": "API-VALIDATION-READINESS: validation.readiness",
        "description": "Contract-only endpoint mapped to `ValidationApplicationService.readiness`. Sprint 67 implements this route in the local FastAPI MVP.",
        "tags": [
          "validation"
        ],
        "x-devpilot-api-id": "API-VALIDATION-READINESS",
        "x-devpilot-operation": "validation.readiness",
        "x-devpilot-domain-service": "ValidationApplicationService.readiness",
        "x-devpilot-status": "implemented-initial",
        "x-devpilot-side-effect": "report_when_adapter_requests_it",
        "x-devpilot-dry-run-default": true,
        "x-devpilot-auth": "token-planned-sprint-68",
        "x-devpilot-policy": "Report writes explicit only",
        "x-devpilot-command-equivalent": "python -m devpilot_core readiness-check --strict --json",
        "responses": {
          "200": {
            "description": "ApplicationResponse wrapping CommandResult.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request envelope or unsupported parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Policy/approval/security gate blocked the operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Contract validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          },
          "500": {
            "description": "Controlled internal error preserving response envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApplicationResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ApplicationRequest"
                  }
                ],
                "properties": {
                  "operation": {
                    "const": "validation.readiness"
                  }
                }
              },
              "example": {
                "operation": "validation.readiness",
                "payload": {},
                "client": "api-local",
                "dry_run": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Finding": {
        "type": "object",
        "required": [
          "id",
          "message",
          "severity"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "fail",
              "block",
              "error"
            ]
          },
          "path": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
      "ApplicationRequest": {
        "type": "object",
        "required": [
          "operation",
          "payload",
          "client",
          "dry_run"
        ],
        "properties": {
          "operation": {
            "type": "string"
          },
          "payload": {
            "type": "object",
            "additionalProperties": true
          },
          "client": {
            "type": "string",
            "default": "api-local"
          },
          "dry_run": {
            "type": "boolean",
            "default": true
          }
        },
        "additionalProperties": true
      },
      "ApplicationResponse": {
        "type": "object",
        "required": [
          "contract",
          "schema_version",
          "operation",
          "ok",
          "exit_code",
          "message",
          "data",
          "findings",
          "generated_at"
        ],
        "properties": {
          "contract": {
            "type": "string",
            "const": "DevPilotApplicationResponse"
          },
          "schema_version": {
            "type": "string"
          },
          "operation": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          },
          "exit_code": {
            "type": "integer",
            "enum": [
              0,
              1,
              2,
              3
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Finding"
            }
          },
          "generated_at": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "ErrorApplicationResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApplicationResponse"
          }
        ],
        "description": "Error envelope preserving ApplicationResponse semantics. Non-2xx responses must still carry operation, exit_code and findings."
      }
    },
    "securitySchemes": {
      "LocalTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-DevPilot-Local-Token",
        "description": "Planned for FUNC-SPRINT-68. Not enforced by Sprint 66 because no HTTP server exists."
      }
    }
  }
}